// packages/usfm-editor-app/src/docs-ui-types.ts
/** Palette / gutter / block-menu icon bucket for Basic & Medium modes. */
export type DocsMenuIconKey =
  | 'paragraph'
  | 'heading'
  | 'poetry'
  | 'list'
  | 'blank'
  | 'indent'
  | 'meta'
  | 'book'
  | 'insert_book'
  | 'insert_verse'
  | 'insert_chapter'
  | 'split_block'
  | 'translator';